Build against Bazel's tensorflow.
This project needs to compile and link against tensorflow-cpu. Before,
the tensorflow-cpu pip package had to be installed manually (e.g. in the
devcontainer) but now Bazel manages it as part of the build system.

This project now uses Bazel's hermetic python interpreter instead of
relying on what is installed on the host system. This prevents python
version mismatches which are a problem because cpython's ABI changes
across minor versions.